+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changes that happened in releases
 * Make sure that singular collection typed attributes are mapped as list in GraphQL
 * Rewrite entity view annotation processor to avoid concurrency issues
 * Fix CTE inlining in set operands
+* Fix custom entity view class bytecode analysis when using 2 stack slot types
 
 ### Backwards-incompatible changes